Move depositLog to types
